submodules/Shremote/requirements.txt
asn1crypto==1.2.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
idna==2.8
paramiko==2.6.0
pycparser==2.19
PyNaCl==1.3.0
PyYAML==5.2
six==1.13.0
